# out/aspnetcore/src/Org.OpenAPITools/Controllers/MinitwitApi.cs
/*
 * Minitwit API
 *
 * Minitwit API
 *
 * The version of the OpenAPI document: 1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class MinitwitApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Get list of users followed by the given user.  - Query param &#x60;?no&#x3D;&#x60; limits result count. - Optionally updates a &#39;latest&#39; global value via &#x60;?latest&#x3D;&#x60; query param.</remarks>
        /// <param name="username"></param>
        /// <param name="authorization">Authorization string of the form &#x60;Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh&#x60;. Used to authenticate as simulator</param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <param name="no">Optional: &#x60;no&#x60; limits result count</param>
        /// <response code="200">Success</response>
        /// <response code="403">Unauthorized - Must include correct Authorization header</response>
        /// <response code="404">User not found (no response body)</response>
        [HttpGet]
        [Route("/fllws/{username}")]
        [ValidateModelState]
        [SwaggerOperation("GetFollow")]
        [SwaggerResponse(statusCode: 200, type: typeof(FollowsResponse), description: "Success")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Unauthorized - Must include correct Authorization header")]
        public abstract IActionResult GetFollow([FromRoute (Name = "username")][Required]string username, [FromHeader (Name = "Authorization")][Required()]string authorization, [FromQuery (Name = "latest")]int? latest, [FromQuery (Name = "no")]int? no);

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Returns the latest ID saved</remarks>
        /// <response code="200">Success</response>
        /// <response code="500">Internal Server Error</response>
        [HttpGet]
        [Route("/latest")]
        [ValidateModelState]
        [SwaggerOperation("GetLatestValue")]
        [SwaggerResponse(statusCode: 200, type: typeof(LatestValue), description: "Success")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error")]
        public abstract IActionResult GetLatestValue();

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Get recent messages.  - Filters out flagged messages - Returns a list of recent messages (max defined by &#x60;?no&#x3D;&#x60; param) - Optionally updates a &#39;latest&#39; global value via &#x60;?latest&#x3D;&#x60; query param.</remarks>
        /// <param name="authorization">Authorization string of the form &#x60;Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh&#x60;. Used to authenticate as simulator</param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <param name="no">Optional: &#x60;no&#x60; limits result count</param>
        /// <response code="200">Success</response>
        /// <response code="403">Unauthorized - Must include correct Authorization header</response>
        [HttpGet]
        [Route("/msgs")]
        [ValidateModelState]
        [SwaggerOperation("GetMessages")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Message>), description: "Success")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Unauthorized - Must include correct Authorization header")]
        public abstract IActionResult GetMessages([FromHeader (Name = "Authorization")][Required()]string authorization, [FromQuery (Name = "latest")]int? latest, [FromQuery (Name = "no")]int? no);

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Get messages for a specific user.  - Returns messages authored by the specified user - Filtered by unflagged - Returns a list of recent messages for the user (max defined by &#x60;?no&#x3D;&#x60; param) - Optionally updates a &#39;latest&#39; global value via &#x60;?latest&#x3D;&#x60; query param.</remarks>
        /// <param name="username"></param>
        /// <param name="authorization">Authorization string of the form &#x60;Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh&#x60;. Used to authenticate as simulator</param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <param name="no">Optional: &#x60;no&#x60; limits result count</param>
        /// <response code="200">Success</response>
        /// <response code="403">Unauthorized - Must include correct Authorization header</response>
        /// <response code="404">User not found (no response body)</response>
        [HttpGet]
        [Route("/msgs/{username}")]
        [ValidateModelState]
        [SwaggerOperation("GetMessagesPerUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Message>), description: "Success")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Unauthorized - Must include correct Authorization header")]
        public abstract IActionResult GetMessagesPerUser([FromRoute (Name = "username")][Required]string username, [FromHeader (Name = "Authorization")][Required()]string authorization, [FromQuery (Name = "latest")]int? latest, [FromQuery (Name = "no")]int? no);

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Follow or unfollow a user on behalf of &#x60;username&#x60;.  - Body must contain either &#x60;follow: &lt;user&gt;&#x60; or &#x60;unfollow: &lt;user&gt;&#x60;</remarks>
        /// <param name="username"></param>
        /// <param name="authorization">Authorization string of the form &#x60;Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh&#x60;. Used to authenticate as simulator</param>
        /// <param name="payload"></param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <response code="204">No Content</response>
        /// <response code="403">Unauthorized - Must include correct Authorization header</response>
        /// <response code="404">User not found (no response body)</response>
        [HttpPost]
        [Route("/fllws/{username}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostFollow")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Unauthorized - Must include correct Authorization header")]
        public abstract IActionResult PostFollow([FromRoute (Name = "username")][Required]string username, [FromHeader (Name = "Authorization")][Required()]string authorization, [FromBody]FollowAction payload, [FromQuery (Name = "latest")]int? latest);

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Post a new message as a specific user.  - Message must include &#x60;content&#x60; in the body - Stored with timestamp and &#x60;flagged&#x3D;0&#x60; - Returns empty body on success - Optionally updates a &#39;latest&#39; global value via &#x60;?latest&#x3D;&#x60; query param.</remarks>
        /// <param name="username"></param>
        /// <param name="authorization">Authorization string of the form &#x60;Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh&#x60;. Used to authenticate as simulator</param>
        /// <param name="payload"></param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <response code="204">No Content</response>
        /// <response code="403">Unauthorized - Must include correct Authorization header</response>
        [HttpPost]
        [Route("/msgs/{username}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostMessagesPerUser")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Unauthorized - Must include correct Authorization header")]
        public abstract IActionResult PostMessagesPerUser([FromRoute (Name = "username")][Required]string username, [FromHeader (Name = "Authorization")][Required()]string authorization, [FromBody]PostMessage payload, [FromQuery (Name = "latest")]int? latest);

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Register a new user. - Optionally updates a &#39;latest&#39; global value via &#x60;?latest&#x3D;&#x60; query param.</remarks>
        /// <param name="payload"></param>
        /// <param name="latest">Optional: &#x60;latest&#x60; value to update</param>
        /// <response code="204">No Content</response>
        /// <response code="400">Bad Request | Possible reasons:  - missing username  - invalid email  - password missing  - username already taken</response>
        [HttpPost]
        [Route("/register")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostRegister")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "Bad Request | Possible reasons:  - missing username  - invalid email  - password missing  - username already taken")]
        public abstract IActionResult PostRegister([FromBody]RegisterRequest payload, [FromQuery (Name = "latest")]int? latest);
    }
}