# out/aspnetcore/src/Org.OpenAPITools/Models/FollowAction.cs
/*
 * Minitwit API
 *
 * Minitwit API
 *
 * The version of the OpenAPI document: 1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Org.OpenAPITools.Converters;

namespace Org.OpenAPITools.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class FollowAction : IEquatable<FollowAction>
    {
        /// <summary>
        /// Username to follow (optional, either this or \&quot;unfollow\&quot;)
        /// </summary>
        /// <value>Username to follow (optional, either this or \&quot;unfollow\&quot;)</value>
        /* <example>Helge</example> */
        [DataMember(Name="follow", EmitDefaultValue=false)]
        public string Follow { get; set; }

        /// <summary>
        /// Username to unfollow (optional, either this or \&quot;follow\&quot;)
        /// </summary>
        /// <value>Username to unfollow (optional, either this or \&quot;follow\&quot;)</value>
        /* <example>Helge</example> */
        [DataMember(Name="unfollow", EmitDefaultValue=false)]
        public string Unfollow { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class FollowAction {\n");
            sb.Append("  Follow: ").Append(Follow).Append("\n");
            sb.Append("  Unfollow: ").Append(Unfollow).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((FollowAction)obj);
        }

        /// <summary>
        /// Returns true if FollowAction instances are equal
        /// </summary>
        /// <param name="other">Instance of FollowAction to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(FollowAction other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Follow == other.Follow ||
                    Follow != null &&
                    Follow.Equals(other.Follow)
                ) && 
                (
                    Unfollow == other.Unfollow ||
                    Unfollow != null &&
                    Unfollow.Equals(other.Unfollow)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Follow != null)
                    hashCode = hashCode * 59 + Follow.GetHashCode();
                    if (Unfollow != null)
                    hashCode = hashCode * 59 + Unfollow.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(FollowAction left, FollowAction right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(FollowAction left, FollowAction right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}